requirements.txt
streamlit
langchain
langchain-aws
langchain-community
langchain-core
boto3
chromadb
duckduckgo-search
pypdf
python-dotenv
sentence-transformers
langchain-huggingface
mcp>=1.10.0
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
requests